Keep Tags within KubeEntityDetailFooter
