+++ packages/manager/src/features/Kubernetes/KubernetesClusterDetail/KubeEntityDetailFooter.tsx
@@ -3,15 +3,19 @@ import { Box, CircleProgress, StyledLinkButton } from '@linode/ui';
 import { pluralize } from '@linode/utilities';
 import Grid from '@mui/material/Grid';
 import { useTheme } from '@mui/material/styles';
+import { enqueueSnackbar } from 'notistack';
 import * as React from 'react';
 
 import { Link } from 'src/components/Link';
+import { TagCell } from 'src/components/TagCell/TagCell';
 import {
   StyledLabelBox,
   StyledListItem,
   sxLastListItem,
   sxListItemFirstChild,
 } from 'src/features/Linodes/LinodeEntityDetail.styles';
+import { useKubernetesClusterMutation } from 'src/queries/kubernetes';
+import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import { formatDate } from 'src/utilities/formatDate';
 
 import { useIsLkeEnterpriseEnabled } from '../kubeUtils';
@@ -40,6 +44,9 @@ export const KubeEntityDetailFooter = React.memo((props: FooterProps) => {
   const {
     aclData,
     clusterCreated,
+    clusterLabel,
+    clusterTags,
+    areClusterLinodesReadOnly,
     clusterId,
     clusterUpdated,
     isLoadingKubernetesACL,
@@ -64,105 +71,154 @@ export const KubeEntityDetailFooter = React.memo((props: FooterProps) => {
     isLkeEnterprisePhase2FeatureEnabled
   );
 
+  const { mutateAsync: updateKubernetesCluster } =
+    useKubernetesClusterMutation(clusterId);
+
+  const handleUpdateTags = React.useCallback(
+    (newTags: string[]) => {
+      return updateKubernetesCluster({
+        tags: newTags,
+      }).catch((e) =>
+        enqueueSnackbar(
+          getAPIErrorOrDefault(e, 'Error updating tags')[0].reason,
+          {
+            variant: 'error',
+          }
+        )
+      );
+    },
+    [updateKubernetesCluster, enqueueSnackbar]
+  );
+
   return (
     <Grid
       container
       data-qa-kube-entity-footer
-      direction="row"
-      spacing={2}
+      direction="column"
       sx={{
-        alignItems: 'center',
-        flex: 1,
-        justifyContent: 'space-between',
-        padding: 0,
+        paddingTop: theme.spacingFunction(8),
       }}
     >
       <Grid
+        container
+        direction="row"
+        spacing={2}
+        sx={{
+          alignItems: 'center',
+          flex: 1,
+          justifyContent: 'space-between',
+          padding: 0,
+        }}
+      >
+        <Grid
+          size={{
+            lg: 'auto',
+            xs: 12,
+          }}
+          sx={{
+            alignItems: 'center',
+            display: 'flex',
+            padding: 0,
+
+            [theme.breakpoints.down('lg')]: {
+              padding: '8px',
+            },
+
+            [theme.breakpoints.between('md', 'xs')]: {
+              display: 'grid',
+              gridTemplateColumns: '50% 2fr',
+            },
+          }}
+        >
+          <Box
+            sx={{
+              display: 'flex',
+              flexWrap: 'wrap',
+              rowGap: 1,
+            }}
+          >
+            <StyledListItem sx={{ ...sxListItemFirstChild }}>
+              <StyledLabelBox component="span">Cluster ID:</StyledLabelBox>{' '}
+              {clusterId}
+            </StyledListItem>
+            {isLkeEnterprisePhase2FeatureEnabled && vpc && (
+              <StyledListItem
+                sx={{
+                  alignItems: 'center',
+                }}
+              >
+                <StyledLabelBox component="span">VPC: </StyledLabelBox>{' '}
+                <Link
+                  data-testid="assigned-lke-cluster-label"
+                  to={`/vpcs/${vpcId}`}
+                >
+                  {vpc?.label ?? `${vpcId}`}
+                </Link>
+                &nbsp;
+                {vpcId && vpc?.label ? `(ID: ${vpcId})` : undefined}
+              </StyledListItem>
+            )}
+            <StyledListItem>
+              <StyledLabelBox component="span">
+                Control Plane ACL:{' '}
+              </StyledLabelBox>{' '}
+              {isLoadingKubernetesACL ? (
+                <Box component="span" sx={{ paddingLeft: 1 }}>
+                  <CircleProgress noPadding size="sm" />
+                </Box>
+              ) : (
+                <StyledLinkButton
+                  disabled={isClusterReadOnly}
+                  onClick={() => setControlPlaneACLDrawerOpen(true)}
+                  sx={(theme) => ({
+                    '&:disabled': {
+                      '& g': {
+                        stroke:
+                          theme.tokens.alias.Content.Icon.Primary.Disabled,
+                      },
+                      color: theme.tokens.alias.Content.Text.Primary.Disabled,
+                    },
+                  })}
+                >
+                  {buttonCopyACL}
+                </StyledLinkButton>
+              )}
+            </StyledListItem>
+            <StyledListItem>
+              <StyledLabelBox component="span">Created:</StyledLabelBox>{' '}
+              {formatDate(clusterCreated, {
+                timezone: profile?.timezone,
+              })}
+            </StyledListItem>
+            <StyledListItem sx={{ ...sxLastListItem }}>
+              <StyledLabelBox component="span">Updated:</StyledLabelBox>{' '}
+              {formatDate(clusterUpdated, {
+                timezone: profile?.timezone,
+              })}
+            </StyledListItem>
+          </Box>
+        </Grid>
+      </Grid>
+      <Grid
         size={{
-          lg: 'auto',
+          lg: 12,
           xs: 12,
         }}
         sx={{
-          alignItems: 'center',
-          display: 'flex',
-          padding: 0,
-
-          [theme.breakpoints.down('lg')]: {
-            padding: '8px',
-          },
-
-          [theme.breakpoints.between('md', 'xs')]: {
-            display: 'grid',
-            gridTemplateColumns: '50% 2fr',
-          },
+          marginLeft: 'auto',
+          marginTop: '8px',
         }}
       >
-        <Box
+        <TagCell
+          disabled={areClusterLinodesReadOnly}
+          entityLabel={clusterLabel}
           sx={{
-            display: 'flex',
-            flexWrap: 'wrap',
-            rowGap: 1,
+            width: '100%',
           }}
-        >
-          <StyledListItem sx={{ ...sxListItemFirstChild }}>
-            <StyledLabelBox component="span">Cluster ID:</StyledLabelBox>{' '}
-            {clusterId}
-          </StyledListItem>
-          {isLkeEnterprisePhase2FeatureEnabled && vpc && (
-            <StyledListItem
-              sx={{
-                alignItems: 'center',
-              }}
-            >
-              <StyledLabelBox component="span">VPC: </StyledLabelBox>{' '}
-              <Link
-                data-testid="assigned-lke-cluster-label"
-                to={`/vpcs/${vpcId}`}
-              >
-                {vpc?.label ?? `${vpcId}`}
-              </Link>
-              &nbsp;
-              {vpcId && vpc?.label ? `(ID: ${vpcId})` : undefined}
-            </StyledListItem>
-          )}
-          <StyledListItem>
-            <StyledLabelBox component="span">
-              Control Plane ACL:{' '}
-            </StyledLabelBox>{' '}
-            {isLoadingKubernetesACL ? (
-              <Box component="span" sx={{ paddingLeft: 1 }}>
-                <CircleProgress noPadding size="sm" />
-              </Box>
-            ) : (
-              <StyledLinkButton
-                disabled={isClusterReadOnly}
-                onClick={() => setControlPlaneACLDrawerOpen(true)}
-                sx={(theme) => ({
-                  '&:disabled': {
-                    '& g': {
-                      stroke: theme.tokens.alias.Content.Icon.Primary.Disabled,
-                    },
-                    color: theme.tokens.alias.Content.Text.Primary.Disabled,
-                  },
-                })}
-              >
-                {buttonCopyACL}
-              </StyledLinkButton>
-            )}
-          </StyledListItem>
-          <StyledListItem>
-            <StyledLabelBox component="span">Created:</StyledLabelBox>{' '}
-            {formatDate(clusterCreated, {
-              timezone: profile?.timezone,
-            })}
-          </StyledListItem>
-          <StyledListItem sx={{ ...sxLastListItem }}>
-            <StyledLabelBox component="span">Updated:</StyledLabelBox>{' '}
-            {formatDate(clusterUpdated, {
-              timezone: profile?.timezone,
-            })}
-          </StyledListItem>
-        </Box>
+          tags={clusterTags}
+          updateTags={handleUpdateTags}
+          view="inline"
+        />
       </Grid>
     </Grid>
   );
